pyproject.toml
[tool.poetry]
name = "hillel-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
pillow = "^11.2.1"
django-modeltranslation = "^0.19.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
